LeetCode-Solution-191: Number of 1 Bits
[Basic Idea]
n & (n - 1) will make the right most 1 in binary representation of n
to be 0. So we check how many times do you need to do this operation
to make n zero. That is actually equals to the numbers of 1 bits in n.

[Complexity]
Time: O(n), where n is the length of bits to represent an integer.