LeetCode-Solution-19: Remove Nth Node From End of List
[Basic Idea]
Let a pioneer pointer goes n-1 steps first, then let a chaser pointer
go through the list step by step from head along with pioneer and
a postChaser with goes one step behind the chaser.
When the pioneer reaches the end of the list, remove the chaser since
we already saved the postChaser. If the chaser is pointing to head,
then just let the head points to the next node.

[Complexity]
Time: O(n)