LeetCode-Solution-90: Subsets II
[Basic Idea]
Each number can be added to the list or not. But for duplications,
we sort the array and make sure that only when previous duplicate
number has been added to the list, then current duplicate number
can be added to the list. E.g, in array [1, 1, 2], only when the
first 1 has been added, the the second 1 can be added. This logic
make sure that the number of duplications can only increase where
decrease will lead to duplicate combinations.

[Complexity]
O(2^n)